Feature/add calculation function

+++ src/main/java/com/obaid/calculator/Calculator.java
@@ -6,12 +6,11 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class Calculator {
-
     public double add(double valueA, double valueB) {
     return valueA + valueB;
     }
 
-    public double subtract(int valueA, int valueB) {
+    public double subtract(double valueA, double valueB) {
         return valueA - valueB;
     }
 }
